Code simplification
- Reasons didn't *need* to be case classes; now they can be instantiated without a factory
- Added "ReasonSet" in preparation for multi-value explanations
- Moved makeRepair methods from CTExplainer to Repair.scala.  It's worth considering whether "Repair"s should be treated as an independent module, or more directly integrated into the Model API.
