+++ src/main/scala/mimir/ctables/CTExplainer.scala
@@ -228,7 +228,7 @@ class CTExplainer(db: Database) extends LazyLogging {
 	{
 		logger.trace(s"GETTING REASONS: $expr")
 		expr match {
-			case v: VGTerm => Map(v.model.name -> makeReason(v, v.args.map(Eval.eval(_,tuple))))
+			case v: VGTerm => Map(v.model.name -> new Reason(v.model, v.idx, v.args.map(Eval.eval(_,tuple))))
 
 			case Conditional(c, t, e) =>
 				(
@@ -300,30 +300,5 @@ class CTExplainer(db: Database) extends LazyLogging {
 		val tuple = finalSchema.map(_._1).zip(baseData.head).toMap
 
 		(tuple, columnExprs, rowCondition)
-	}
-
-	def makeReason(term: VGTerm, v: Seq[PrimitiveValue]): Reason =
-		makeReason(term.model, term.idx, v)
-	def makeReason(model: Model, idx: Int, v: Seq[PrimitiveValue]): Reason =
-	{
-    Reason(
-      model.reason(idx, v),
-      model.name,
-      idx,
-      v,
-      makeRepair(model, idx, v)
-    )
-	}
-
-	def makeRepair(term: VGTerm, v: Seq[PrimitiveValue]): Repair =
-		makeRepair(term.model, term.idx, v)
-	def makeRepair(model: Model, idx: Int, v: Seq[PrimitiveValue]): Repair =
-	{
-		model match {
-			case finite:( Model with FiniteDiscreteDomain ) =>
-				RepairFromList(finite.getDomain(idx, v))
-			case _ => 
-				RepairByType(model.varType(idx, v.map(_.getType)))
-		}
 	}
 }
